fix: LEAP-1651: Properly update meta text for regions
Meta text is saved in regions by Normalization mixin and it's
serialized into every result of this region. When we update the meta text
we should get this value again from region, not previously stored one in result.

So the priority of sources for meta value was changed.
Plus "control meta" was removed.